+++ src/lambda/trips/TripsEvent.ts
